Split delete and connect node tests
